Reorganized components/views and started work on Login component
This commit contains several large changes but most importantly has initial workdone on the Login View component. At the time of commit, the component is able to make a valid XMLHttpRequest to the backend api at api.transfr.info through CORS. Further work is need for Vuex store to serve as single source of truth.