Added in basic User view
This commit starts work on the User view and is able to populate data from the backend. Note that this commit also the abandons the use of Vuex for state management. Considering the complexity that Vuex brings, it is overkill for this proejct and not worth the trouble of properly configuring the store object. May wind up selecting a different implementation of Flux given the frustation on using Vuex