Added blur focus event to complex input
this commit adds the blur event handler for complex input fields. This is as iusers do not intuitively click the plus button or hit enter on the field. This results in the user not saving data when they are ready to submit the changes. The behavior is now when the user changes focus onto another field, the data is now added