made custom address input and added custom presentation for addresses
This commit adds HTML to handle the editting of address values. This is a change
reflected in the schema to reflect the additional information that is typically expected in the address field, i.e. to be standards compliant. Subsequently,
additional styling and presentation changes were needed for the proper represent-tation of the backing data.