Added focus events across event to replicate similar behavior to prior attributes
This commit adds focus events on all elements to allow user to focus on element and adds an address. Not quite the same as the focusing on the background does not add the address but is very close. Needed to be done this was as blur always returned the body element