Adds vuex-presistedstate plugin to presist state between refreshes; resolved blank load
This commit adds a vuex-presistedstate plugin to presist state between page
refreshes. This is necessary as vuex state is not by default. In the case
where the user does refresh the page, state would be cleared and result in
undefined data reaching render and causing failure. Adding this plugin and
presisting state to sessionStorage between refreshes alleviates this issue.

this commit adds the ability for vuex to load data from api. Do have blank cards on load, probably related to creativity issue
and props. Will now refactor CardManager not needing to use props anymore.