Able to add LabeledInput values to store; refactored store structure
This commit is able to now add new values using the blank fields in the
LabeledInput store. Note that the component now interates over not
the computed property of values anymore but rather the return value
of a method. This seems to be actually reactive and response to changes
in computed values that derived from the Vuex store. It's a tad unorthodox
but the Vue instance seems not to complain

Also, this refactors the Vuex card module to have getters, setters, etc., as
their own modules now as the file was starting to grow a little unwieldy.