Implemented Removal of LabeledInput values from Vuex
This commit implements the removal of values from labeled input from
the Vuex store. Unfortunately, in order to make it work, I had to
use forceUpdate which is less than ideal. For some reason, regardless
of what method used, Vue.delete, splice, use of computed property or
watcher, I can't get reactivity to function here as optimal.

Maybe the usage of the method is not great, or something changes are
getting buffered in the async update queue. The latter seems likely
considering changes to the new value fields trigger a update of the
list