Started implementation of JSON Patch creation from Vuex store
This commit implements basic functionality of creating a JSON
PATCH operations from Vuex store. This for now only compares
the values where a simple string is stored as the value.

Work needs to be done for combined values like address and email
Also logic needs to be considered to implement extraneous patches
and edits from hitting the API. Might not be necessary due to
how the diffing is explicitly done instead of enumerating over
all possible patches created in response to events