Implementing sorting of patch operations of attributes with array values
This commits implements sorting of JSON patch operations created from
enumerating over the values of array values of relevant card attributes.
This is as the spec for JSON PATCH dictates that operations are executed
in the order they are inserted in the body.

Since array values are manipulated either by full replacement or operated
over by index, a premature removal or addition may result in a failed patch.
As such, the output of the generatePatchesOverArray is sorted so that replace
operations happen first, then remove ops, and finally add operations. This
ordering is expected to lower the chance of the aforementioned scenario
from happening.