Implemented Vuex Patch action to API; started work modifying resetCard mutation
This commit primarily implements the actual patching and subsequent
merging of changes from the API back into the Vuex store. This is
accomplished by comparing the original state and the modified state
as in the store to generate a patch object, and then sends the patch to the API.

Upon successfully patch of the card, the vuex action then call the mutation to
assign and update the values in the store as shown in the backup and updates
the newly updated state as the new unmodified, original state. Work is still
needed to clean up temporary values (though it might be deemed as unnecessary).

Additionally, as the original state now contains a denormalized card state,
some modifications were started for the resetting of store state via the
resetCard mutation.