Remove erroneous state parameter being passed in
This commit remotes the extra 'state' argument that was passed into
generateDiffForPatch. For the record, it was not variable hoisting
or accidential reactivity that was causing the issue, it was passing
the state object for the module as the original parameter that was causing the the issue.
In the future, probably should look at ways at stopping this, mechanisms to catch additional
parameters being passed than expected